updated db seed

+++ database_seed.sql
@@ -64,8 +64,9 @@ create table order_item
     user_id   int          not null,
     item_id   int          not null,
     item_name varchar(255) not null,
-    cost      int          not null,
+    cost      float(11, 2) not null,
     qty       int          not null,
     primary key (order_id, user_id, item_id)
 );
+
